refactor: data 디렉터리를 database로 수정
